Remove Dispatch timer from subscribeToViewModel

+++ RickAndMorty-iOS/Views/Episodes/EpisodesViewController.swift
@@ -24,6 +24,7 @@ class EpisodesViewController: UIViewController {
         super.viewDidLoad()
 
         configureDataSource()
+        showEmptyData()
         subscribeToViewModel()
         viewModel.currentPage = 1
     }
@@ -37,25 +38,26 @@ class EpisodesViewController: UIViewController {
         episodesView.collectionView.refreshControl?.addTarget(self, action: #selector(onRefresh), for: .valueChanged)
     }
 
-    func subscribeToViewModel() {
+    func showEmptyData() {
         snapshot.deleteAllItems()
         snapshot.appendSections([.appearance, .empty])
         snapshot.appendItems([], toSection: .appearance)
         snapshot.appendItems(Array(repeatingExpression: EmptyData(id: UUID()), count: 8), toSection: .empty)
-        self.dataSource.apply(snapshot, animatingDifferences: false)
+        self.dataSource.apply(snapshot, animatingDifferences: true)
+    }
+
+    func subscribeToViewModel() {
         viewModel.episodes.sink(receiveValue: { [self] episodes in
-            DispatchQueue.main.asyncAfter(deadline: .now() + 1) { [self] in
+            if !episodes.isEmpty {
                 snapshot.deleteAllItems()
                 snapshot.appendSections([.appearance])
                 snapshot.appendItems(episodes, toSection: .appearance)
                 self.dataSource.apply(snapshot, animatingDifferences: true)
             }
-
             // Dismiss refresh control.
             DispatchQueue.main.async {
                 self.episodesView.collectionView.refreshControl?.endRefreshing()
             }
-
         }).store(in: &cancellables)
     }
 
@@ -76,8 +78,8 @@ extension EpisodesViewController {
             // section with empty episode cells
             if indexPath.section == 1 {
                 let cell = collectionView.dequeueConfiguredReusableCell(using: self.episodesView.episodeCell, for: indexPath, item: data as? EmptyData )
-                // add loading effect
-                showLoading(currentCell: cell)
+//                showLoadingAnimation(currentCell: cell)
+//                hideLoadingAnimation(currentCell: cell)
                 return cell
             }
